+++ src/index.html
@@ -47,8 +47,9 @@
     <section id="section2" name="bosque">
       <div class="bosque" id="bosque">
         <h1 class="title">Hechizos</h1>
-        <i class="fa-solid fa-magnifying-glass"></i>
-        <input type="text" placeholder="search" id="search-wand">
+        <button class="search-button"><i class="fa-solid fa-magnifying-glass"></i></button>
+        <input type="search" placeholder="search" id="search-wand">
+
         <!-- Filtrado de informacion-->
         <select id="informacion">
           <option value="">Todos </option>
@@ -66,31 +67,10 @@
         </select>
         <div id="contenedorspells">
         </div>
-      </div>
-
-      <template data-user-template>
-        <div class="box-grid" id="box-grid">
-          <div class="box">
-            <p></p>
-          </div>
-      </template>
-
-
-                    <!-- class="spells-container -->
-                    <ul  id="container-ul">
-
-
-                    </ul>
-                        <li>
-                    <template id="spells">
-                            
-                    </template>
-                        </li>
-                </div>
-            </section>
-                
+        <div class="box-grid" id="box-grid"></div>
 
     </section>
+
 
     <div class="box-grid" id="box-grid">
       <!-- Información mostrada en pantalla-->
@@ -106,11 +86,9 @@
       <div class="all-books">
         <div class="book">
           <div class="front-cover"></div>
-          <!--
-                        <div class="page"></div>
-            <div class="page"></div>
-                    <div class="page"></div>
-                    -->
+          <div class="page"></div>
+          <div class="page"></div>
+          <div class="page"></div>
           <div class="last-page">
             <img src="../img/Harry Potter.jpg" alt="Harry
                   Potter">
@@ -133,20 +111,16 @@
       <h1>Casas u otro</h1>
     </section>
 
-            <section id="section5">
-                <div>
-                    <h1>Casas u otro</h1>
-                </div>
-            </section>
-        </div>
-    </main>
+    </div>
 
-    <footer>
-        <span>Diseñado por </span>
-        <a href="https://www.linkedin.com/in/martihuaman/" target="_blank" rel="noopener noreferrer">
-            Karen Huamán</a> y <a href="#">Adriana Armijos</a>
-    </footer>
-    <script src="main.js" type="module" defer></script>
+  </main>
+  <footer>
+    <span>Diseñado por </span>
+    <a href="https://www.linkedin.com/in/martihuaman/" target="_blank" rel="noopener noreferrer">
+      Karen Huamán</a> y <a href="https://www.linkedin.com/in/adriana-armijos-721385185/">Adriana Armijos</a>
+  </footer>
+
+  <script src="main.js" type="module" defer></script>
 </body>
 
 </html>
